rename pkg mem -> memory
